+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.4.6 November 18, 2024
 
 - New #125: Add `subscription_period` and `business_connection_id` parameters to `createInvoiceLink` method.
+- New #125: Add `subscription_expiration_date`, `is_recurring` and `is_first_recurring` fields to `SuccessfulPayment`
+  type.
 
 ## 0.4.5 October 31, 2024